improved microphone support
Some devices (I dont know why) say that they are support a specific sample rate but when we try to record something with this sample rate our app crashes. Now we check if the samplerate is really supported and if not we try a list of samplerates until we found one tha
t is supported. If we dont find any we show the user an error message and report this to crashalytics