implement #8 - Show absolute Minimum/Maximum in Recording Details
It makes sense to show the real min and max values also on the recording details view, because it says you wich voicepitch you can reach with training. For this I added a getMin() and getMax() mehtod to the pitch calculator and displayed the results on the recording play fragment.